rav1e::api::internal

Struct InterConfig

pub struct InterConfig {
    reorder: bool,
    pub(crate) multiref: bool,
    pub(crate) pyramid_depth: u64,
    pub(crate) group_input_len: u64,
    group_output_len: u64,
    pub(crate) switch_frame_interval: u64,
}

The set of options that controls frame re-ordering and reference picture selection. The options stored here are invariant over the whole encode.

Fields
reorder: bool

Whether frame re-ordering is enabled.

multiref: bool

Whether P-frames can use multiple references.

pyramid_depth: u64

The depth of the re-ordering pyramid. The current code cannot support values larger than 2.

group_input_len: u64

Number of input frames in group.

group_output_len: u64

Number of output frames in group. This includes both hidden frames and โshow existing frameโ frames.

switch_frame_interval: u64

Interval between consecutive S-frames. Keyframes reset this interval. This MUST be a multiple of group_input_len.

Implementations

impl InterConfig

pub(crate) fn new(enc_config: &EncoderConfig) -> InterConfig

pub(crate) fn get_idx_in_group_output(&self, output_frameno_in_gop: u64) -> u64

Get the index of an output frame in its re-ordering group given the output frame number of the frame in the current keyframe gop. When re-ordering is disabled, this always returns 0.

pub(crate) fn get_order_hint( &self, output_frameno_in_gop: u64, idx_in_group_output: u64, ) -> u32

Get the order-hint of an output frame given the output frame number of the frame in the current keyframe gop and the index of that output frame in its re-ordering gorup.

pub(crate) const fn get_level(&self, idx_in_group_output: u64) -> u64

Get the level of the current frame in the pyramid.

pub(crate) const fn get_slot_idx(&self, level: u64, order_hint: u32) -> u32

pub(crate) const fn get_show_frame(&self, idx_in_group_output: u64) -> bool

pub(crate) const fn get_show_existing_frame( &self, idx_in_group_output: u64, ) -> bool

pub(crate) fn get_input_frameno( &self, output_frameno_in_gop: u64, gop_input_frameno_start: u64, ) -> u64

const fn max_reordering_latency(&self) -> u64

pub(crate) const fn keyframe_lookahead_distance(&self) -> u64

pub(crate) const fn allowed_ref_frames(&self) -> &[RefType]