rav1e

Module config

Encoder configuration and settings

Structs

Config

Contains the encoder configuration.

EncoderConfig

Encoder settings which impact the produced bitstream.

GrainTableSegment

Specifies parameters for enabling decoder-side grain synthesis for a segment of video from start_time to end_time.

NoiseGenArgs

Settings and video data defining how to generate the film grain params.

RateControlConfig

Rate control configuration

RateControlSummary

Rate control pass summary

SpeedSettings

Contains the speed settings.

Enums

CpuFeatureLevel

InvalidConfig

Enumeration of possible invalid configuration errors.

PredictionModesSetting

Prediction modes to search.

RateControlError

Rate control errors

TransferFunction

Constants

NUM_UV_COEFFS

The max number of coefficients per chroma plane for grain synthesis

NUM_UV_POINTS

The max number of scaling points per chroma plane for grain synthesis

NUM_Y_COEFFS

The max number of luma coefficients for grain synthesis

NUM_Y_POINTS

The max number of luma scaling points for grain synthesis