rav1e::dist

Module rust

Constants

AREA_DIVISORS 🔒

Lookup table for 2^AREA_DIVISOR_BITS / (1 + x)

AREA_DIVISOR_BITS 🔒

Number of bits of precision used in AREA_DIVISORS

GET_WEIGHTED_SSE_SHIFT

Number of bits rounded off before summing in get_weighted_sse

Functions

butterfly 🔒

cdef_dist_kernel

Computes a distortion metric of the sum of squares weighted by activity. w and h should be <= 8.

get_sad

Compute the sum of absolute differences over a block. w and h can be at most 128, the size of the largest block.

get_satd

Sum of absolute transformed differences over a block. w and h can be at most 128, the size of the largest block. Use the sum of 4x4 and 8x8 hadamard transforms for the transform, but revert to sad on edges when these transforms do not fit into w and h. 4x4 transforms instead of 8x8 transforms when width or height < 8.

get_weighted_sse

Computes weighted sum of squared error.

hadamard2d 🔒

hadamard4_1d 🔒

hadamard4x4 🔒 ⚠

hadamard8_1d 🔒

hadamard8x8 🔒 ⚠