rav1e::ec

Struct WriterBase

pub struct WriterBase<S> {
    rng: u16,
    cnt: i16,
    fake_bits_frac: u32,
    s: S,
}

Fields
rng: u16

The number of values in the current range.

cnt: i16

The number of bits of data in the current value.

fake_bits_frac: u32

Extra offset added to tell() and tell_frac() to approximate costs of actually coding a symbol

s: S

Use-specific storage

Implementations

impl<S> WriterBase<S>

A few local helper functions needed by the Writer that are not part of the public interface.

const fn new(storage: S) -> Self

Internal constructor called by the subtypes that implement the actual encoder and Recorder.

const fn lr_compute(&self, fl: u16, fh: u16, nms: u16) -> (u32, u16)

Compute low and range values from token cdf values and local state

fn frac_compute(nbits_total: u32, rng: u32) -> u32

Given the current total integer number of bits used and the current value of rng, computes the fraction number of bits used to OD_BITRES precision. This is used by od_ec_enc_tell_frac() and od_ec_dec_tell_frac(). nbits_total: The number of whole bits currently used, i.e., the value returned by od_ec_enc_tell() or od_ec_dec_tell(). rng: The current value of rng from either the encoder or decoder state. Return: The number of bits scaled by 2**OD_BITRES. This will always be slightly larger than the exact value (e.g., all rounding error is in the positive direction).

const fn recenter(r: u32, v: u32) -> u32

impl WriterBase<WriterRecorder>

Replay implementation specific to the Recorder

pub fn replay(&mut self, dest: &mut dyn StorageBackend)

Replays the partially-computed range tokens out of the Recorder’s storage and into the passed in Writer, which may be an Encoder or another Recorder. Clears the Recorder after replay.

impl WriterBase<WriterEncoder>

Done implementation specific to the Encoder

pub fn done(&mut self) -> Vec<u8>

Indicates that there are no more symbols to encode. Flushes remaining state into coding and returns a vector containing the final bitstream.