rav1e::ec

Trait StorageBackend

pub trait StorageBackend {
    // Required methods
    fn store(&mut self, fl: u16, fh: u16, nms: u16);
    fn stream_bits(&mut self) -> usize;
    fn checkpoint(&mut self) -> WriterCheckpoint;
    fn rollback(&mut self, _: &WriterCheckpoint);
}

StorageBackend is an internal trait used to tie a specific Writer implementation’s storage to the generic Writer. It would be private, but Rust is deprecating ‘private trait in a public interface’ support.

Required Methods

fn store(&mut self, fl: u16, fh: u16, nms: u16)

Store partially-computed range code into given storage backend

fn stream_bits(&mut self) -> usize

Return bit-length of encoded stream to date

fn checkpoint(&mut self) -> WriterCheckpoint

Backend implementation of checkpoint to pass through Writer interface

fn rollback(&mut self, _: &WriterCheckpoint)

Backend implementation of rollback to pass through Writer interface