rav1e

Module encoder

Structs

CodedFrameData

These frame invariants are only used on coded frames, i.e. non-SEFs. They are stored separately to avoid useless allocations when we do not need them.

DeblockState

FrameInvariants

FrameState

ReferenceFrame

ReferenceFramesSet

SBSQueueEntry

SegmentationState

Sequence

Enums

CDEFSearchMethod

Tune

Constants

DELTA_FRAME_ID_LENGTH 🔒

FRAME_ID_LENGTH 🔒

IMPORTANCE_BLOCK_SIZE

Size of blocks for the importance computation, in pixels.

MAX_NUM_OPERATING_POINTS 🔒

MAX_NUM_SPATIAL_LAYERS 🔒

MAX_NUM_TEMPORAL_LAYERS 🔒

Statics

TEMPORAL_DELIMITER

Functions

build_raw_tile_group 🔒

check_lf_queue 🔒

diff 🔒

Write into dst the difference between the blocks at src1 and src2

encode_block_post_cdef

Panics

encode_block_pre_cdef

encode_block_with_modes

encode_frame

Panics

encode_partition_bottomup 🔒

encode_partition_topdown 🔒

encode_show_existing_frame

Write a packet containing only the placeholder that tells the decoder to present the already decoded frame present at frame_to_show_map_idx

encode_tile 🔒

encode_tile_group 🔒

encode_tx_block

For a transform block, predict, transform, quantize, write coefficients to a bitstream, dequantize, inverse-transform.

get_initial_cdfcontext 🔒

get_initial_segmentation 🔒

get_qidx 🔒

motion_compensate

Panics

poly2 🔒

pos_to_lvl 🔒

save_block_motion

update_rec_buffer

write_key_frame_obus 🔒

write_temporal_delimiter

Errors

write_tile_group_header 🔒

write_tx_blocks

Panics

write_tx_tree