rav1e::encoder

Struct FrameInvariants

pub struct FrameInvariants<T: Pixel> {

    pub sequence: Arc<Sequence>,
    pub config: Arc<EncoderConfig>,
    pub width: usize,
    pub height: usize,
    pub render_width: u32,
    pub render_height: u32,
    pub frame_size_override_flag: bool,
    pub render_and_frame_size_different: bool,
    pub sb_width: usize,
    pub sb_height: usize,
    pub w_in_b: usize,
    pub h_in_b: usize,
    pub input_frameno: u64,
    pub order_hint: u32,
    pub show_frame: bool,
    pub showable_frame: bool,
    pub error_resilient: bool,
    pub intra_only: bool,
    pub allow_high_precision_mv: bool,
    pub frame_type: FrameType,
    pub frame_to_show_map_idx: u32,
    pub use_reduced_tx_set: bool,
    pub reference_mode: ReferenceMode,
    pub use_prev_frame_mvs: bool,
    pub partition_range: PartitionRange,
    pub globalmv_transformation_type: [GlobalMVMode; 7],
    pub num_tg: usize,
    pub large_scale_tile: bool,
    pub disable_cdf_update: bool,
    pub allow_screen_content_tools: u32,
    pub force_integer_mv: u32,
    pub primary_ref_frame: u32,
    pub refresh_frame_flags: u32,
    pub allow_intrabc: bool,
    pub use_ref_frame_mvs: bool,
    pub is_filter_switchable: bool,
    pub is_motion_mode_switchable: bool,
    pub disable_frame_end_update_cdf: bool,
    pub allow_warped_motion: bool,
    pub cdef_search_method: CDEFSearchMethod,
    pub cdef_damping: u8,
    pub cdef_bits: u8,
    pub cdef_y_strengths: [u8; 8],
    pub cdef_uv_strengths: [u8; 8],
    pub delta_q_present: bool,
    pub ref_frames: [u8; 7],
    pub ref_frame_sign_bias: [bool; 7],
    pub rec_buffer: ReferenceFramesSet<T>,
    pub base_q_idx: u8,
    pub dc_delta_q: [i8; 3],
    pub ac_delta_q: [i8; 3],
    pub lambda: f64,
    pub me_lambda: f64,
    pub dist_scale: [DistortionScale; 3],
    pub me_range_scale: u8,
    pub use_tx_domain_distortion: bool,
    pub use_tx_domain_rate: bool,
    pub idx_in_group_output: u64,
    pub pyramid_level: u64,
    pub enable_early_exit: bool,
    pub tx_mode_select: bool,
    pub enable_inter_txfm_split: bool,
    pub default_filter: FilterMode,
    pub enable_segmentation: bool,
    pub t35_metadata: Box<[T35]>,
    pub cpu_feature_level: CpuFeatureLevel,
    pub coded_frame_data: Option<CodedFrameData<T>>,

}

Fields
sequence: Arc<Sequence>config: Arc<EncoderConfig>width: usizeheight: usizerender_width: u32render_height: u32frame_size_override_flag: boolrender_and_frame_size_different: boolsb_width: usizesb_height: usizew_in_b: usizeh_in_b: usizeinput_frameno: u64order_hint: u32show_frame: boolshowable_frame: boolerror_resilient: boolintra_only: boolallow_high_precision_mv: boolframe_type: FrameTypeframe_to_show_map_idx: u32use_reduced_tx_set: boolreference_mode: ReferenceModeuse_prev_frame_mvs: boolpartition_range: PartitionRangeglobalmv_transformation_type: [GlobalMVMode; 7]num_tg: usizelarge_scale_tile: booldisable_cdf_update: boolallow_screen_content_tools: u32force_integer_mv: u32primary_ref_frame: u32refresh_frame_flags: u32allow_intrabc: booluse_ref_frame_mvs: boolis_filter_switchable: boolis_motion_mode_switchable: booldisable_frame_end_update_cdf: boolallow_warped_motion: boolcdef_search_method: CDEFSearchMethodcdef_damping: u8cdef_bits: u8cdef_y_strengths: [u8; 8]cdef_uv_strengths: [u8; 8]delta_q_present: boolref_frames: [u8; 7]ref_frame_sign_bias: [bool; 7]rec_buffer: ReferenceFramesSet<T>base_q_idx: u8dc_delta_q: [i8; 3]ac_delta_q: [i8; 3]lambda: f64me_lambda: f64dist_scale: [DistortionScale; 3]me_range_scale: u8use_tx_domain_distortion: booluse_tx_domain_rate: boolidx_in_group_output: u64pyramid_level: u64enable_early_exit: booltx_mode_select: boolenable_inter_txfm_split: booldefault_filter: FilterModeenable_segmentation: boolt35_metadata: Box<[T35]>cpu_feature_level: CpuFeatureLevel

Target CPU feature level.

coded_frame_data: Option<CodedFrameData<T>>
Implementations

impl<T: Pixel> FrameInvariants<T>

pub fn new(config: Arc<EncoderConfig>, sequence: Arc<Sequence>) -> Self

Panics

If the size of T does not match the sequence’s bit depth

pub fn new_key_frame( config: Arc<EncoderConfig>, sequence: Arc<Sequence>, gop_input_frameno_start: u64, t35_metadata: Box<[T35]>, ) -> Self

pub(crate) fn new_inter_frame( previous_coded_fi: &Self, inter_cfg: &InterConfig, gop_input_frameno_start: u64, output_frameno_in_gop: u64, next_keyframe_input_frameno: u64, error_resilient: bool, t35_metadata: Box<[T35]>, ) -> Option<Self>

Returns the created FrameInvariants, or None if this should be a placeholder frame.

pub fn is_show_existing_frame(&self) -> bool

pub fn clone_without_coded_data(&self) -> Self

pub fn set_ref_frame_sign_bias(&mut self)

pub fn get_frame_subtype(&self) -> usize

fn pick_strength_from_q(&mut self, qps: &QuantizerParameters)

pub fn set_quantizers(&mut self, qps: &QuantizerParameters)

pub fn sb_size_log2(&self) -> usize

pub fn film_grain_params(&self) -> Option<&GrainTableSegment>

pub fn frame_timestamp(&self) -> u64