rav1e

Module me

Macros

search_pattern 🔒

Declares an array of motion vectors in structure of arrays syntax. Compared to [search_pattern_subpel], this version creates motion vectors in fullpel resolution (x8).

search_pattern_subpel 🔒

Declares an array of motion vectors in structure of arrays syntax.

Structs

FrameMEStats

MEStats

MVCandidateRD

Holds data from computing rate distortion of a motion vector.

MotionEstimationSubsets 🔒

MotionSearchResult

Result of motion search.

Enums

MVSamplingMode

Constants

DIAMOND_R1_PATTERN 🔒

Diamond pattern of radius 1 as shown below. Unlike DIAMOND_R1_PATTERN, the vectors have been shifted fullpel scale.

DIAMOND_R1_PATTERN_SUBPEL 🔒

Diamond pattern of radius 1 as shown below. For fullpel search, use DIAMOND_R1_PATTERN_FULLPEL since it has been scaled for fullpel search.

HEXAGON_PATTERN 🔒

A hexagon pattern around a center point. The pattern is ordered so that the offsets circle around the center. This is done to allow pruning locations covered by the last iteration.

SQUARE_REFINE_PATTERN 🔒

A small square pattern around a center point.

UMH_PATTERN 🔒

Uneven multi-hexagon search pattern around a center point. Used for locating irregular movement.

Functions

compute_mv_rd 🔒

Compute the rate distortion stats for a motion vector.

estimate_motion

estimate_sb_motion 🔒

estimate_tile_motion

full_pixel_me 🔒

full_search 🔒

fullpel_diamond_search 🔒

Run a full pixel diamond search. The search is run on multiple step sizes.

get_best_predictor 🔒

get_fullpel_mv_rd 🔒

get_mv_range 🔒

get_mv_rate 🔒

get_subpel_mv_rd 🔒

get_subset_predictors 🔒

hexagon_search 🔒

Perform hexagon search and refine afterwards.

refine_subsampled_motion_estimate 🔒

Refine motion estimation that was computed one level of subsampling up.

refine_subsampled_sb_motion 🔒

save_me_stats 🔒

sub_pixel_me 🔒

subpel_diamond_search 🔒

Run a subpixel diamond search. The search is run on multiple step sizes.

uneven_multi_hex_search 🔒

Perform an uneven multi-hexagon search. There are 4 stages:

Type Aliases

ReadGuardMEStats

cbindgen:ignore

RefMEStats

cbindgen:ignore

WriteGuardMEStats

cbindgen:ignore