rav1e

Module prelude

Commonly used types and traits.

Modules

color

Color model information

config

Encoder Configuration

context

Encoder Context

v_frame

Structs

ChromaticityPoint

Chromaticity coordinates as defined by CIE 1931, expressed as 0.16 fixed-point values.

ColorDescription

Signal the content color description

Config

Contains the encoder configuration.

ContentLight

High dynamic range content light level

Context

The encoder context.

DiffGenerator

EncoderConfig

Encoder settings which impact the produced bitstream.

Frame

Represents a raw video frame

FrameParameters

Optional per-frame encoder parameters

GrainTableSegment

Specifies parameters for enabling decoder-side grain synthesis for a segment of video from start_time to end_time.

MasteringDisplay

High dynamic range mastering display color volume

MotionSpeedSettings

Speed settings related to motion estimation and motion vector selection

NoiseGenArgs

Settings and video data defining how to generate the film grain params.

Opaque

Opaque type to be passed from Frame to Packet

Packet

Represents a packet.

PartitionRange

Range of block sizes to use.

PartitionSpeedSettings

Speed settings related to partition decision

Plane

One data plane of a frame.

PlaneConfig

Plane-specific configuration.

PredictionSpeedSettings

Speed settings related to intra prediction mode selection

RateControlConfig

Rate control configuration

RateControlSummary

Rate control pass summary

Rational

A rational number.

Sequence

SpeedSettings

Contains the speed settings.

T35

A single T.35 metadata packet.

TilingInfo

Tiling information

TransformSpeedSettings

Speed settings related to transform size and type decision

Enums

BlockSize

ChromaSamplePosition

Sample position for subsampled chroma

ChromaSampling

Chroma subsampling format

ColorPrimaries

Supported Color Primaries

EncoderStatus

Status that can be returned by Context functions.

FrameType

Possible types of a frame.

FrameTypeOverride

Override the frame type decision

InvalidConfig

Enumeration of possible invalid configuration errors.

MatrixCoefficients

Matrix coefficients

PixelRange

Allowed pixel value range

PixelType

Types that can be used as pixel types.

PredictionMode

PredictionModesSetting

Prediction modes to search.

RateControlError

Rate control errors

RcData

Rate Control Data

SGRComplexityLevel

Search level for self guided restoration

SceneDetectionSpeed

Prediction modes to search.

SegmentationLevel

Search level for segmentation

TransferCharacteristics

Supported Transfer Characteristics

TransferFunction

Tune

TxType

Constants

DEFAULT_GRAIN_SEED

A randomly generated u16 to be used as a starting random seed for grain synthesis. The idea behind using a constant random seed is so that encodes are deterministic and reproducible.

NUM_UV_COEFFS

The max number of coefficients per chroma plane for grain synthesis

NUM_UV_POINTS

The max number of scaling points per chroma plane for grain synthesis

NUM_Y_COEFFS

The max number of luma coefficients for grain synthesis

NUM_Y_POINTS

The max number of luma scaling points for grain synthesis

Traits

CastFromPrimitive

Trait for casting between primitive types.

IntoFrame

Types which can be converted into frames.

Pixel

A type that can be used as a pixel type.

Functions

generate_photon_noise_params

Generates a set of photon noise parameters for a segment of video given a set of args.

parse_grain_table

This file has the implementation details of the grain table.

write_grain_table

Write a set of generated film grain params to a table file, using the standard film grain table format supported by aomenc, rav1e, and svt-av1.

Type Aliases

ScalingPoints