rav1e::prelude::color

Enum ColorPrimaries

#[repr(C)]
pub enum ColorPrimaries {
    BT709 = 1,
    Unspecified = 2,
    BT470M = 4,
    BT470BG = 5,
    BT601 = 6,
    SMPTE240 = 7,
    GenericFilm = 8,
    BT2020 = 9,
    XYZ = 10,
    SMPTE431 = 11,
    SMPTE432 = 12,
    EBU3213 = 22,
}

Supported Color Primaries

As defined by “Color primaries” section of ISO/IEC 23091-4/ITU-T H.273

Variants

BT709 = 1

BT.709

Unspecified = 2

Unspecified, must be signaled or inferred outside of the bitstream

BT470M = 4

BT.470 System M (historical)

BT470BG = 5

BT.470 System B, G (historical)

BT601 = 6

BT.601-7 525 (SMPTE 170 M)

SMPTE240 = 7

SMPTE 240M (historical)

GenericFilm = 8

Generic film

BT2020 = 9

BT.2020, BT.2100

XYZ = 10

SMPTE 248 (CIE 1921 XYZ)

SMPTE431 = 11

SMPTE RP 431-2

SMPTE432 = 12

SMPTE EG 432-1

EBU3213 = 22

EBU Tech. 3213-E

Implementations

impl ColorPrimaries

pub fn variants() -> [&'static str; 12]

Returns an array of valid values which can be converted into this enum.

pub fn descriptions() -> [(&'static [&'static str], &'static [&'static str]); 12]

Returns an array of touples (variants, description)