rav1e::prelude::color

Enum TransferCharacteristics

#[repr(C)]
pub enum TransferCharacteristics {

    BT709 = 1,
    Unspecified = 2,
    BT470M = 4,
    BT470BG = 5,
    BT601 = 6,
    SMPTE240 = 7,
    Linear = 8,
    Log100 = 9,
    Log100Sqrt10 = 10,
    IEC61966 = 11,
    BT1361 = 12,
    SRGB = 13,
    BT2020_10Bit = 14,
    BT2020_12Bit = 15,
    SMPTE2084 = 16,
    SMPTE428 = 17,
    HLG = 18,

}

Supported Transfer Characteristics

As defined by “Transfer characteristics” section of ISO/IEC 23091-4/ITU-TH.273.

Variants

BT709 = 1

BT.709

Unspecified = 2

Unspecified, must be signaled or inferred outside of the bitstream

BT470M = 4

BT.470 System M (historical)

BT470BG = 5

BT.470 System B, G (historical)

BT601 = 6

BT.601-7 525 (SMPTE 170 M)

SMPTE240 = 7

SMPTE 240 M

Linear = 8

Linear

Log100 = 9

Logarithmic (100:1 range)

Log100Sqrt10 = 10

Logarithmic ((100 * √10):1 range)

IEC61966 = 11

IEC 61966-2-4

BT1361 = 12

BT.1361 extended color gamut system (historical)

SRGB = 13

sRGB or sYCC

BT2020_10Bit = 14

BT.2020 10-bit systems

BT2020_12Bit = 15

BT.2020 12-bit systems

SMPTE2084 = 16

SMPTE ST 2084, ITU BT.2100 PQ

SMPTE428 = 17

SMPTE ST 428

HLG = 18

BT.2100 HLG (Hybrid Log Gamma), ARIB STD-B67

Implementations

impl TransferCharacteristics

pub fn variants() -> [&'static str; 17]

Returns an array of valid values which can be converted into this enum.

pub fn descriptions() -> [(&'static [&'static str], &'static [&'static str]); 17]

Returns an array of touples (variants, description)