rav1e::prelude

Enum BlockSize

pub enum BlockSize {

    BLOCK_4X4,
    BLOCK_4X8,
    BLOCK_8X4,
    BLOCK_8X8,
    BLOCK_8X16,
    BLOCK_16X8,
    BLOCK_16X16,
    BLOCK_16X32,
    BLOCK_32X16,
    BLOCK_32X32,
    BLOCK_32X64,
    BLOCK_64X32,
    BLOCK_64X64,
    BLOCK_64X128,
    BLOCK_128X64,
    BLOCK_128X128,
    BLOCK_4X16,
    BLOCK_16X4,
    BLOCK_8X32,
    BLOCK_32X8,
    BLOCK_16X64,
    BLOCK_64X16,

}

Variants

BLOCK_4X4

BLOCK_4X8

BLOCK_8X4

BLOCK_8X8

BLOCK_8X16

BLOCK_16X8

BLOCK_16X16

BLOCK_16X32

BLOCK_32X16

BLOCK_32X32

BLOCK_32X64

BLOCK_64X32

BLOCK_64X64

BLOCK_64X128

BLOCK_128X64

BLOCK_128X128

BLOCK_4X16

BLOCK_16X4

BLOCK_8X32

BLOCK_32X8

BLOCK_16X64

BLOCK_64X16

Implementations

impl BlockSize

pub const BLOCK_SIZES_ALL: usize = 22usize

pub const BLOCK_SIZES: usize = 16usize

pub fn from_width_and_height_opt( w: usize, h: usize, ) -> Result<BlockSize, InvalidBlockSize>

Errors

Returns InvalidBlockSize if the given w and h do not produce a valid block size.

pub fn from_width_and_height(w: usize, h: usize) -> BlockSize

Panics

If the given w and h do not produce a valid block size.

pub fn cfl_allowed(self) -> bool

pub const fn width(self) -> usize

pub const fn area(self) -> usize

width * height

pub const fn width_log2(self) -> usize

pub const fn width_mi_log2(self) -> usize

pub const fn width_mi(self) -> usize

pub fn width_imp_b(self) -> usize

pub const fn height(self) -> usize

pub const fn height_log2(self) -> usize

pub const fn height_mi_log2(self) -> usize

pub const fn height_mi(self) -> usize

pub fn height_imp_b(self) -> usize

pub const fn tx_size(self) -> TxSize

pub const fn subsampled_size( self, xdec: usize, ydec: usize, ) -> Result<BlockSize, InvalidBlockSize>

Source: Subsampled_Size (AV1 specification section 5.11.38)

Errors

Returns InvalidBlockSize if the given block size cannot be subsampled in the requested way.

pub fn largest_chroma_tx_size(self, xdec: usize, ydec: usize) -> TxSize

Panics

Will panic if the subsampling is not possible

pub const fn is_sqr(self) -> bool

pub const fn is_sub8x8(self, xdec: usize, ydec: usize) -> bool

pub const fn sub8x8_offset(self, xdec: usize, ydec: usize) -> (isize, isize)

pub const fn subsize( self, partition: PartitionType, ) -> Result<BlockSize, InvalidBlockSize>

Errors

Returns InvalidBlockSize if the block size cannot be split in the requested way.

pub const fn is_rect_tx_allowed(self) -> bool