rav1e::prelude

Struct Sequence

pub struct Sequence {

    pub profile: u8,
    pub num_bits_width: u32,
    pub num_bits_height: u32,
    pub bit_depth: usize,
    pub chroma_sampling: ChromaSampling,
    pub chroma_sample_position: ChromaSamplePosition,
    pub pixel_range: PixelRange,
    pub color_description: Option<ColorDescription>,
    pub mastering_display: Option<MasteringDisplay>,
    pub content_light: Option<ContentLight>,
    pub max_frame_width: u32,
    pub max_frame_height: u32,
    pub frame_id_numbers_present_flag: bool,
    pub frame_id_length: u32,
    pub delta_frame_id_length: u32,
    pub use_128x128_superblock: bool,
    pub order_hint_bits_minus_1: u32,
    pub force_screen_content_tools: u32,
    pub force_integer_mv: u32,
    pub still_picture: bool,
    pub reduced_still_picture_hdr: bool,
    pub enable_filter_intra: bool,
    pub enable_intra_edge_filter: bool,
    pub enable_interintra_compound: bool,
    pub enable_masked_compound: bool,
    pub enable_dual_filter: bool,
    pub enable_order_hint: bool,
    pub enable_jnt_comp: bool,
    pub enable_ref_frame_mvs: bool,
    pub enable_warped_motion: bool,
    pub enable_superres: bool,
    pub enable_cdef: bool,
    pub enable_restoration: bool,
    pub enable_large_lru: bool,
    pub enable_delayed_loopfilter_rdo: bool,
    pub operating_points_cnt_minus_1: usize,
    pub operating_point_idc: [u16; 32],
    pub display_model_info_present_flag: bool,
    pub decoder_model_info_present_flag: bool,
    pub level_idx: [u8; 32],
    pub tier: [usize; 32],
    pub film_grain_params_present: bool,
    pub timing_info_present: bool,
    pub tiling: TilingInfo,
    pub time_base: Rational,

}

Fields
profile: u8

OBU Sequence header of AV1

num_bits_width: u32num_bits_height: u32bit_depth: usizechroma_sampling: ChromaSamplingchroma_sample_position: ChromaSamplePositionpixel_range: PixelRangecolor_description: Option<ColorDescription>mastering_display: Option<MasteringDisplay>content_light: Option<ContentLight>max_frame_width: u32max_frame_height: u32frame_id_numbers_present_flag: boolframe_id_length: u32delta_frame_id_length: u32use_128x128_superblock: boolorder_hint_bits_minus_1: u32force_screen_content_tools: u32

0 - force off 1 - force on 2 - adaptive

force_integer_mv: u32

0 - Not to force. MV can be in 1/4 or 1/8 1 - force to integer 2 - adaptive

still_picture: bool

Video is a single frame still picture

reduced_still_picture_hdr: bool

Use reduced header for still picture

enable_filter_intra: bool

enables/disables filter_intra

enable_intra_edge_filter: bool

enables/disables corner/edge filtering and upsampling

enable_interintra_compound: bool

enables/disables interintra_compound

enable_masked_compound: bool

enables/disables masked compound

enable_dual_filter: bool

0 - disable dual interpolation filter 1 - enable vert/horiz filter selection

enable_order_hint: bool

0 - disable order hint, and related tools jnt_comp, ref_frame_mvs, frame_sign_bias if 0, enable_jnt_comp and enable_ref_frame_mvs must be set zs 0.

enable_jnt_comp: bool

0 - disable joint compound modes 1 - enable it

enable_ref_frame_mvs: bool

0 - disable ref frame mvs 1 - enable it

enable_warped_motion: bool

0 - disable warped motion for sequence 1 - enable it for the sequence

enable_superres: bool

0 - Disable superres for the sequence, and disable transmitting per-frame superres enabled flag. 1 - Enable superres for the sequence, and also enable per-frame flag to denote if superres is enabled for that frame.

enable_cdef: bool

To turn on/off CDEF

enable_restoration: bool

To turn on/off loop restoration

enable_large_lru: bool

To turn on/off larger-than-superblock loop restoration units

enable_delayed_loopfilter_rdo: bool

allow encoder to delay loop filter RDO/coding until after frame reconstruciton is complete

operating_points_cnt_minus_1: usizeoperating_point_idc: [u16; 32]display_model_info_present_flag: booldecoder_model_info_present_flag: boollevel_idx: [u8; 32]tier: [usize; 32]

seq_tier in the spec. One bit: 0 or 1.

film_grain_params_present: booltiming_info_present: booltiling: TilingInfotime_base: Rational
Implementations

impl Sequence

pub fn new(config: &EncoderConfig) -> Sequence

Panics

Panics if the resulting tile sizes would be too large.

pub const fn get_relative_dist(&self, a: u32, b: u32) -> i32

pub fn get_skip_mode_allowed<T: Pixel>( &self, fi: &FrameInvariants<T>, inter_cfg: &InterConfig, reference_select: bool, ) -> bool