rav1e::prelude

Trait IntoFrame

pub trait IntoFrame<T: Pixel> {
    // Required method
    fn into(self) -> (Option<Arc<Frame<T>>>, Option<FrameParameters>);
}

Types which can be converted into frames.

This trait is used in Context::send_frame to allow for passing in frames with optional frame parameters and optionally frames wrapped in Arc (to allow for zero-copy, since the encoder uses frames in Arc internally).

Required Methods

fn into(self) -> (Option<Arc<Frame<T>>>, Option<FrameParameters>)

Converts the type into a tuple of frame and parameters.