rav1e

Module rdo

Structs

Distortion

DistortionScale

Fixed point arithmetic version of distortion scale

PartitionGroupParameters

PartitionParameters

RawDistortion

ScaledDistortion

Enums

RDOType

Functions

cdef_dist_wxh

clip_visible_bsize

compute_distortion 🔒

compute_rd_cost

compute_tx_distortion 🔒

distortion_scale

Compute a scaling factor to multiply the distortion of a block by, this factor is determined using temporal RDO.

distortion_scale_for

dmv_in_range 🔒

estimate_rate

get_sub_partitions

inter_frame_rdo_mode_decision 🔒

intra_frame_rdo_mode_decision 🔒

luma_chroma_mode_rdo 🔒

rdo_cfl_alpha

Panics

rdo_loop_decision

Passed in a superblock offset representing the upper left corner of the LRU area we’re optimizing. This area covers the largest LRU in any of the present planes, but may consist of a number of superblocks and full, smaller LRUs in the other planes

rdo_loop_plane_error 🔒

rdo_mode_decision

RDO-based mode decision

rdo_partition_decision

RDO-based single level partitioning decision

rdo_partition_none 🔒

rdo_partition_simple 🔒

rdo_tx_size_type

rdo_tx_type_decision

RDO-based transform type decision If cw_checkpoint is None, a checkpoint for cw’s (ContextWriter) current state is created and stored for later use.

spatiotemporal_scale

Panics

sse_wxh

Sum of Squared Error for a wxh block Currently limited to w and h of valid blocks