rav1e::rdo

Function distortion_scale

pub fn distortion_scale<T: Pixel>(
    fi: &FrameInvariants<T>,
    frame_bo: PlaneBlockOffset,
    bsize: BlockSize,
) -> DistortionScale

Compute a scaling factor to multiply the distortion of a block by, this factor is determined using temporal RDO.

Panics

If called with bsize of 8x8 or smaller

If the coded frame data doesn’t exist on the FrameInvariants