rav1e::rdo

Function rdo_tx_type_decision

pub fn rdo_tx_type_decision<T: Pixel>(
    fi: &FrameInvariants<T>,
    ts: &mut TileStateMut<'_, T>,
    cw: &mut ContextWriter<'_>,
    cw_checkpoint: &mut Option<ContextWriterCheckpoint>,
    mode: PredictionMode,
    ref_frames: [RefType; 2],
    mvs: [MotionVector; 2],
    bsize: BlockSize,
    tile_bo: TileBlockOffset,
    tx_size: TxSize,
    tx_set: TxSet,
    tx_types: &[TxType],
    cur_best_rd: f64,
) -> (TxType, f64)

RDO-based transform type decision If cw_checkpoint is None, a checkpoint for cw’s (ContextWriter) current state is created and stored for later use.

Panics

If a writer checkpoint is never created before or within the function. This should never happen and indicates a development error.

If the best RD found is negative. This should never happen and indicates a development error.