Module rav1e::prelude::config

Encoder Configuration

Modules

v_frame

Structs

Config

Contains the encoder configuration.

DiffGenerator

EncoderConfig

Encoder settings which impact the produced bitstream.

GrainTableSegment

Specifies parameters for enabling decoder-side grain synthesis for a segment of video from start_time to end_time.

MotionSpeedSettings

Speed settings related to motion estimation and motion vector selection

NoiseGenArgs

Settings and video data defining how to generate the film grain params.

PartitionRange

Range of block sizes to use.

PartitionSpeedSettings

Speed settings related to partition decision

PredictionSpeedSettings

Speed settings related to intra prediction mode selection

RateControlConfig

Rate control configuration

RateControlSummary

Rate control pass summary

SpeedSettings

Contains the speed settings.

TilingInfo

Tiling information

TransformSpeedSettings

Speed settings related to transform size and type decision

Enums

InvalidConfig

Enumeration of possible invalid configuration errors.

PredictionModesSetting

Prediction modes to search.

RateControlError

Rate control errors

SGRComplexityLevel

Search level for self guided restoration

SceneDetectionSpeed

Prediction modes to search.

SegmentationLevel

Search level for segmentation

TransferFunction

Constants

DEFAULT_GRAIN_SEED

A randomly generated u16 to be used as a starting random seed for grain synthesis. The idea behind using a constant random seed is so that encodes are deterministic and reproducible.

NUM_UV_COEFFS

The max number of coefficients per chroma plane for grain synthesis

NUM_UV_POINTS

The max number of scaling points per chroma plane for grain synthesis

NUM_Y_COEFFS

The max number of luma coefficients for grain synthesis

NUM_Y_POINTS

The max number of luma scaling points for grain synthesis

Functions

generate_photon_noise_params

Generates a set of photon noise parameters for a segment of video given a set of args.

parse_grain_table

This file has the implementation details of the grain table.

write_grain_table

Write a set of generated film grain params to a table file, using the standard film grain table format supported by aomenc, rav1e, and svt-av1.

Type Aliases

ScalingPoints