Module rav1e::version

Version information

The information is recovered from Cargo.toml and git describe, when available.

use rav1e::version;
use semver::Version;

let major = version::major();
let minor = version::minor();
let patch = version::patch();

let short = version::short();

let v1 = Version::new(major, minor, patch);
let v2 = Version::parse(&short).unwrap();

assert_eq!(v1.major, v2.major);

Functions

full

Version information with the information provided by git describe --tags.

hash

Commit hash (short)

long

Version information as presented in [package] version followed by the short commit hash if present.

major

Major version component

minor

Minor version component

patch

Patch version component

short

Version information as presented in [package] version.